Django_Study/pyproject.toml
[tool.poetry]
name = "django-study"
version = "0.1.0"
description = "python django project"
authors = ["daeun0608 <glace0608@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
pyjwt = "<2.10.0"
djangorestframework-simplejwt = "^5.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
